fix a number of int size conversion warnings
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@11692 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/MiscUI/FilterEdit.cpp
@@ -376,7 +376,7 @@ void CFilterEdit::DrawDimText()
 	dcDraw.SelectObject((*GetFont()));
 	dcDraw.SetTextColor(GetSysColor(COLOR_GRAYTEXT));
 	dcDraw.SetBkColor(GetSysColor(COLOR_WINDOW));
-	dcDraw.DrawText(m_pCueBanner, _tcslen(m_pCueBanner), &rRect, DT_CENTER | DT_VCENTER);
+	dcDraw.DrawText(m_pCueBanner, (int)_tcslen(m_pCueBanner), &rRect, DT_CENTER | DT_VCENTER);
 	dcDraw.RestoreDC(iState);
 	return;
 }
